Remove django-extensions
* Clean up imports
* Use Python 3 shutil.which
* Stop integration test on build error
* Add custom UUIDField and use it for uuid model attributes
* Remove django_extensions from existing migrations
* Add migration for current changes
* Update requirements

+++ integration/run.sh
@@ -6,6 +6,12 @@ cd ${__dir}
 
 docker compose build archivematica-storage-service
 
+status=$?
+
+if [ $status -ne 0 ]; then
+    exit $status
+fi
+
 docker compose run --rm archivematica-storage-service
 
 status=$?
